Create 5G NETWORK AWARENESS INTERACTIVE POWER BI DASHBOARD
The "5G Network Awareness and Usage" project aims to educate and empower individuals and organizations about the potential of 5G technology. It seeks to enhance infrastructure, develop innovative applications, and address security concerns to maximize the benefits of 5G networks. By promoting awareness and optimizing usage, the project contributes to the efficient utilization of this cutting-edge technology, impacting various sectors and enhancing the overall digital experience.